+++ CHANGELOG.md
@@ -6,13 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.11] - 2022-05-31
+
 ### Added
-- bundle with webpack (vsix >3MB --> ~150KB)!
+- bundle with webpack (vsix >3MB --> ~500KB)!
+- disable save button when there is no active editor
 
 ### Changed
-- disable save button when there is no active editor
 - update font awesome to version 6
-- remove uuid dependency and use node's crypto library instead
+- remove uuid dependency, use node's crypto library
 
 ## [0.1.10] - 2022-05-25
 ### Changed